Moved session fixtures to conftests
